New role system: -added new role moderator -added users page -modified role handling

+++ server/routes/admin.js
@@ -3,18 +3,19 @@ module.exports = (io, logger) => {
         router = express.Router()
 
     const Submition = require("./../models/submition"),
-        VoteElement = require("./../models/voteElement")
+        VoteElement = require("./../models/voteElement"),
+        User = require("./../models/user")
 
     const lyricsClient = require("lyrics-finder")
 
-    // check if user is admin
+    // check if user is admin or moderator
     router.use((req, res, next) => {
-        if (req.user && req.user.role == "admin") next()
+        if (req.user && req.user.role.level > 0) next()
         else res.status(404).render("error")
     })
 
     router.get("/", async (req, res) => {
-        res.render("admin", { submitQueue: await Submition.find({}) })
+        res.render("admin", { submitQueue: await Submition.find({}), user: req.user })
     })
 
     router.put("/verdict", async (req, res) => {
@@ -63,11 +64,73 @@ module.exports = (io, logger) => {
         io.emit("removeVoteElement", id)
     })
 
+    // check if user is admin for the rest of routes
+    router.use((req, res, next) => {
+        if (req.user && req.user.role.level > 1) next()
+        else res.status(404).render("error")
+    })
+
+    const roles = require("./../modules/roles")
+
+    const fieldRenames = {
+        "name": "name",
+        "email": "email",
+        "id": "googleId"
+    }
+
+    const searchMethods = Object.keys(fieldRenames)
+
+    router.get("/users", async (req, res) => {
+        let searchBy = req.query.by,
+            searchQuery = req.query.query
+
+        let field = null
+        if (searchBy && searchQuery) {
+            searchBy = decodeURIComponent(searchBy)
+            searchQuery = decodeURIComponent(searchQuery)
+            field = fieldRenames[searchBy]
+        }
+
+        let filter = {}
+        if (field)
+            filter[field] = { "$regex": searchQuery, "$options": "i" }
+
+        let users = await User.find(filter).limit(10)
+        res.render("users", {
+            searchMethods: searchMethods,
+            users: users,
+            user: req.user,
+            by: searchBy,
+            query: searchQuery
+        })
+    })
+
+    router.put("/users/promote/:id", async (req, res) => {
+        let id = decodeURIComponent(req.params.id)
+        let targetUser = await User.findOne({ googleId: id })
+        if (!req.user.canPromote(targetUser)) return res.status(403).send()
+        let newRole = roles.getNextRole(targetUser.role)
+        logger.info(`${req.user.name}#${req.user.googleId} promoted data ${targetUser.name}#${targetUser.googleId} to ${newRole.name}`)
+        await User.updateOne({ googleId: id }, { $set: { role: newRole.level } })
+        res.send(newRole)
+    })
+
+    router.put("/users/depromote/:id", async (req, res) => {
+        let id = decodeURIComponent(req.params.id)
+        let targetUser = await User.findOne({ googleId: id })
+        if (!req.user.canDepromote(targetUser)) return res.status(403).send()
+        let newRole = roles.getPrevRole(targetUser.role)
+        logger.info(`${req.user.name}#${req.user.googleId} depromoted data ${targetUser.name}#${targetUser.googleId} to ${newRole.name}`)
+        await User.updateOne({ googleId: id }, { $set: { role: newRole.level } })
+        res.send(newRole)
+    })
+
     router.get("/reset", async (req, res) => {
+        if (req.user.level !== Infinity) return res.status(404).render("error")
         logger.info(`${req.user.name}#${req.user.googleId} reseted data`)
         await Submition.deleteMany({})
         await VoteElement.deleteMany({})
-        res.json({ code: "success" })
+        res.redirect("/")
     })
 
     io.on("connection", socket => {
